DSND-1901: Restructure classes into sub-packages
