+++ src/itest/java/uk/gov/companieshouse/appointments/subdelta/Configuration.java
@@ -6,6 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.kafka.test.context.EmbeddedKafka;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.companieshouse.appointments.subdelta.kafka.TestConfig;
 
 @SpringBootTest
 @CucumberContextConfiguration
