+++ src/main/java/uk/gov/companieshouse/appointments/subdelta/companyprofile/AppointmentsClient.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.appointments.subdelta;
+package uk.gov.companieshouse.appointments.subdelta.companyprofile;
 
 import java.util.function.Supplier;
 import org.springframework.stereotype.Component;
