+++ src/main/java/uk/gov/companieshouse/appointments/subdelta/kafka/MessageLoggingAspect.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.appointments.subdelta;
+package uk.gov.companieshouse.appointments.subdelta.kafka;
 
 import static uk.gov.companieshouse.appointments.subdelta.Application.NAMESPACE;
 
@@ -13,6 +13,7 @@ import org.aspectj.lang.annotation.Before;
 import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.messaging.Message;
 import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.appointments.subdelta.kafka.Consumer;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
