+++ src/test/java/uk/gov/companieshouse/appointments/subdelta/companyprofile/ResponseHandlerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.appointments.subdelta;
+package uk.gov.companieshouse.appointments.subdelta.companyprofile;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -11,6 +11,9 @@ import org.junit.jupiter.api.function.Executable;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.error.ApiErrorResponseException;
 import uk.gov.companieshouse.api.handler.exception.URIValidationException;
+import uk.gov.companieshouse.appointments.subdelta.companyprofile.ResponseHandler;
+import uk.gov.companieshouse.appointments.subdelta.exception.NonRetryableException;
+import uk.gov.companieshouse.appointments.subdelta.exception.RetryableException;
 
 @ExtendWith(MockitoExtension.class)
 class ResponseHandlerTest {
