+++ src/test/java/uk/gov/companieshouse/appointments/subdelta/kafka/ConsumerInvalidTopicTest.java
@@ -1,20 +1,14 @@
-package uk.gov.companieshouse.appointments.subdelta;
+package uk.gov.companieshouse.appointments.subdelta.kafka;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.verify;
-import static uk.gov.companieshouse.appointments.subdelta.TestUtils.STREAM_COMPANY_PROFILE_ERROR_TOPIC;
-import static uk.gov.companieshouse.appointments.subdelta.TestUtils.STREAM_COMPANY_PROFILE_INVALID_TOPIC;
-import static uk.gov.companieshouse.appointments.subdelta.TestUtils.STREAM_COMPANY_PROFILE_RETRY_TOPIC;
-import static uk.gov.companieshouse.appointments.subdelta.TestUtils.STREAM_COMPANY_PROFILE_TOPIC;
+import static uk.gov.companieshouse.appointments.subdelta.kafka.TestUtils.STREAM_COMPANY_PROFILE_ERROR_TOPIC;
+import static uk.gov.companieshouse.appointments.subdelta.kafka.TestUtils.STREAM_COMPANY_PROFILE_INVALID_TOPIC;
+import static uk.gov.companieshouse.appointments.subdelta.kafka.TestUtils.STREAM_COMPANY_PROFILE_RETRY_TOPIC;
+import static uk.gov.companieshouse.appointments.subdelta.kafka.TestUtils.STREAM_COMPANY_PROFILE_TOPIC;
 
 import java.io.ByteArrayOutputStream;
-import java.util.Collections;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
+import java.util.concurrent.Future;
 import org.apache.avro.io.DatumWriter;
 import org.apache.avro.io.Encoder;
 import org.apache.avro.io.EncoderFactory;
@@ -23,18 +17,18 @@ import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.clients.producer.RecordMetadata;
+import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.kafka.test.EmbeddedKafkaBroker;
 import org.springframework.kafka.test.context.EmbeddedKafka;
 import org.springframework.kafka.test.utils.KafkaTestUtils;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
-import uk.gov.companieshouse.stream.EventRecord;
-import uk.gov.companieshouse.stream.ResourceChangedData;
+import uk.gov.companieshouse.appointments.subdelta.Application;
 
 @SpringBootTest(classes = Application.class)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
@@ -46,9 +40,9 @@ import uk.gov.companieshouse.stream.ResourceChangedData;
         controlledShutdown = true,
         partitions = 1
 )
-@Import(TestConfig.class)
 @ActiveProfiles("test_main_nonretryable")
-class ConsumerNonRetryableExceptionTest {
+@Import(TestConfig.class)
+class ConsumerInvalidTopicTest {
 
     @Autowired
     private EmbeddedKafkaBroker embeddedKafkaBroker;
@@ -59,35 +53,25 @@ class ConsumerNonRetryableExceptionTest {
     @Autowired
     private KafkaProducer<String, byte[]> testProducer;
 
-    @Autowired
-    private CountDownLatch latch;
-
-    @MockBean
-    private ServiceRouter router;
-
     @Test
-    void testRepublishToCompanyProfileInvalidMessageTopicIfNonRetryableExceptionThrown() throws Exception {
+    void testPublishToCompanyProfileInvalidMessageTopicIfInvalidDataDeserialised() throws Exception {
         //given
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         Encoder encoder = EncoderFactory.get().directBinaryEncoder(outputStream, null);
-        DatumWriter<ResourceChangedData> writer = new ReflectDatumWriter<>(ResourceChangedData.class);
-        writer.write(new ResourceChangedData("", "", "", "", "{}",
-                new EventRecord("", "", Collections.emptyList())), encoder);
+        DatumWriter<String> writer = new ReflectDatumWriter<>(String.class);
+        writer.write("bad data", encoder);
 
         embeddedKafkaBroker.consumeFromAllEmbeddedTopics(testConsumer);
-        doThrow(NonRetryableException.class).when(router).route(any());
 
         //when
-        testProducer.send(
-                new ProducerRecord<>(STREAM_COMPANY_PROFILE_TOPIC, 0, System.currentTimeMillis(),
-                        "key", outputStream.toByteArray()));
-        if (!latch.await(5L, TimeUnit.SECONDS)) {
-            fail("Timed out waiting for latch");
-        }
+        Future<RecordMetadata> future = testProducer.send(
+                new ProducerRecord<>(STREAM_COMPANY_PROFILE_TOPIC,
+                        0, System.currentTimeMillis(), "key", outputStream.toByteArray()));
+        future.get();
         ConsumerRecords<?, ?> consumerRecords = KafkaTestUtils.getRecords(testConsumer, 10000L, 2);
 
         //then
-        assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, STREAM_COMPANY_PROFILE_TOPIC),
+        MatcherAssert.assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, STREAM_COMPANY_PROFILE_TOPIC),
                 is(1));
         assertThat(TestUtils.noOfRecordsForTopic(consumerRecords,
                 STREAM_COMPANY_PROFILE_RETRY_TOPIC), is(0));
@@ -95,6 +79,5 @@ class ConsumerNonRetryableExceptionTest {
                 STREAM_COMPANY_PROFILE_ERROR_TOPIC), is(0));
         assertThat(TestUtils.noOfRecordsForTopic(consumerRecords,
                 STREAM_COMPANY_PROFILE_INVALID_TOPIC), is(1));
-        verify(router).route(any());
     }
 }
