+++ src/test/java/uk/gov/companieshouse/appointments/subdelta/kafka/ResourceChangedDataDeserialiserTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.appointments.subdelta;
+package uk.gov.companieshouse.appointments.subdelta.kafka;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
@@ -19,6 +19,8 @@ import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
+import uk.gov.companieshouse.appointments.subdelta.exception.InvalidPayloadException;
+import uk.gov.companieshouse.appointments.subdelta.kafka.ResourceChangedDataDeserialiser;
 import uk.gov.companieshouse.stream.EventRecord;
 import uk.gov.companieshouse.stream.ResourceChangedData;
